aStarSearch optimizations, repeatedA* progress, documentation updates.
aStarSearch now takes openList and closedList as arguments,
previously instantiated them each time it was run.

repeatedAStarSearch is coming along. A section of the presumedPath
initializer was compartmentized into it's own method for ease of
use during the repeated searches the agent does in the while loop
(called initializePath()).

Documentation was updated for aStarSearch.