Big commit; bug introduced; see description for details.
- Node toString() method now returns x and y coordinates as well as
"Not solvable." if it is not solvable.

- All gridworld objects were changed from Integer[][]-type 2D arrays
to int[][]-type 2D arrays (this was causing issues with initialization
of values).

- The SearchTracker object has been introduced to store information
regarding a current search. Information includes the current position
of the agent, the direction it's traveling (forward or backward),
the current value of the counter variable, and whether the search
is A* repeated or regular A* search.

All methods were updated to use this object instead of long lists
of method parameters.

- A* Repeated search has been implemented. There currently exists a
bug where the program does not finish the A* Repeated Search (making
it run indefinitely). However, it does compute a path for the agent to
travel utilizing the proper logic. The issue lies in getting the
algorithm to stop searching afterwards.

- The moveAgent() helper method was introduced to assist the A*
repeated search algorithm.

- the initializePath() method was renamed to reversePath(), since
that provides a more accurate self-descriptive name.

reversePath() was edited to ensure that no circular linked lists
were accidentally being created from it's logic.

-aStarSearch() was edited so that repeatedAStarSearch() can utilize
it to compute a path from the agent to the goal.

The method now has additional parameters to specify whether or not
it is being used in the repeatedAStarSearch() method.]

-performAction() was modified to change it's behavior based on
whether it's being used in aStarSearch or repeatedAStarSearch.