repeatedAStarSearch fix; debugging statements*; various method fixes
- repeatedAStarSearch now properly functions as expected.

- A new "path" field was added to the SearchTracker object to track
the path the agent travels. Previously the "agent" field was
used to store both the agent's path and the agent's current
position.

- A new "finished" field was added to SearchTracker to keep track
of whether the agent has reached the goal or not during moveAgent.

- repeatedAStarSearch, moveAgent, aStarSearch, and performAction
methods were updated to utilize SearchTracker's "path" and
"finished" fields.

- reversePath was unobfuscated and given some brief documentation.

- performAction was fixed to properly initialize its "goal" variable.

* There are a lot of leftover debugging statements that I'm leaving
in this commit in case I want to go back and do some further testing.
The next commit will have these statements removed and represent a
functionally presentable version of the project.